move migration to separate file
